fix the 80% bug
Summary: Large video files were being downloaded every time you download a clip, this cause it to not get audio in time and eventually timeout. This uses the video in the browser to get audio. Instead of waiting for the file to download, now you have to wait for the whole clip to play through, but it is much faster.

Test Plan: